requirements.txt
azure-identity==1.13.0b4
Flask==2.3.2
openai==0.27.7
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
python-dotenv==1.0.0
pyodbc==4.0.39